refactor: isolate code related to register
